RAC-721: Move public api regarding categories from Structure to Enrichment
