+++ src/Akeneo/Pim/Enrichment/PublicApi/Categories/SqlCountTotalCategoriesPerTree.php
@@ -2,14 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Structure\Bundle\Query\PublicApi\Category\Sql;
+namespace Akeneo\Pim\Enrichment\PublicApi\Categories;
 
-use Akeneo\Pim\Enrichment\Component\Category\Model\Category;
-use Akeneo\Pim\Enrichment\Component\Product\Normalizer\Standard\TranslationNormalizer;
-use Akeneo\Pim\Structure\Component\Query\PublicApi\Category\CategoryTree;
-use Akeneo\Pim\Structure\Component\Query\PublicApi\Category\CountTotalCategoriesPerTree;
-use Akeneo\Pim\Structure\Component\Query\PublicApi\Category\FindCategoryTrees;
-use Akeneo\Tool\Component\Classification\Repository\CategoryRepositoryInterface;
 use Doctrine\DBAL\Connection;
 use Webmozart\Assert\Assert;
