+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/queries.yml
@@ -87,17 +87,6 @@ services:
         arguments:
             - '@database_connection'
 
-    akeneo.pim.structure.query.find_category_trees:
-        class: Akeneo\Pim\Structure\Bundle\Query\PublicApi\Category\Sql\SqlFindCategoryTrees
-        arguments:
-            - '@pim_catalog.repository.category'
-            - '@pim_catalog.normalizer.standard.translation'
-
-    akeneo.pim.structure.query.count_total_categories_per_tree:
-        class: Akeneo\Pim\Structure\Bundle\Query\PublicApi\Category\Sql\SqlCountTotalCategoriesPerTree
-        arguments:
-            - '@database_connection'
-
     akeneo.pim.structure.query.get_family_translations:
         class: Akeneo\Pim\Structure\Bundle\Query\PublicApi\Family\Sql\SqlGetFamilyTranslations
         arguments:
